requirements.txt
asgiref==3.7.2
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==3.2.25
django-mongoengine==0.5.6
dnspython==2.3.0
idna==3.8
mongoengine==0.28.2
oauthlib==3.2.2
paho-mqtt==2.1.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pymongo==4.7.3
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.3.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==4.4.2
social-auth-storage-mongoengine==1.0.1
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==2.0.7
